refactor: 루트, /community 경로 재조정 (#73)
- 루트 경로와 /community를 같은 페이지로.
- 루트 경로에서 자동으로 /community로 이동.
